Only log "Using PATH_TO_raven.conf" message on startup if conf file exists.
Currently we log a message indicating that a raven.conf file is being used
even if one does not exists. This commit changes the logic to only display
this message if a config file exists and logs a separate message
if no config file exists. Additionally, a warning is now logged if the file
path passed in the -conf flag does not exist.